Fix passing empty weight vectors to SE2
The mxIgraph library changed from always returning filled a weight vector even
if unweighted (all 1s) to returning an empty weight vector if unweighted. This
needs to be handled by testing if the weight vector is empty and passing NULL
instead of the vector in that case.

Otherwise SE2 was using uninitialized weight values.